feat: Add services module with core service traits for application architecture
